requirements.txt
numpy==1.11.3
scipy==0.18.1
matplotlib==2.0.0
setuptools==35.0.2
PyYAML==3.12
